Use system Jansson throughout the native implementation
Replace helper and renderer-driver JSON handling with Jansson, remove vendored yyjson, and select nonnegative signed-64-bit protocol values with negative zero accepted. Bound provider counters before serialization and add current installed protocol and output-fault regressions.

Host installed and upstream suites, ASan/UBSan checks, glibc 2.28 build and RPM checks, offline Fedora source rebuild, 400 renderer comparisons and final paired latency gates pass. Retain adoption evidence and keep the superseded prototype in Git history.

+++ native/git.c
@@ -381,7 +381,7 @@ static uint64_t counter(const unsigned char *text, size_t length, char sign)
     for (; i < length; ++i) {
         if (text[i] < '0' || text[i] > '9') return 0;
         unsigned digit = text[i] - '0';
-        if (result > (UINT64_MAX - digit) / 10) return 0;
+        if (result > ((uint64_t)INT64_MAX - digit) / 10) return 0;
         result = result * 10 + digit;
     }
     return result;
